requirements.txt
opencv-python==4.2.0.34
opencv-contrib-python==4.2.0.34
Pillow==7.1.2
h5py==2.10.0
imutils==0.5.3
numpy~=1.19.5

tkinter==8.6

python-telegram-bot-13.6==13.6
openpyxl~=3.0.7
